feat(magifsdp, odin): Refactor MagiFSDP Parameter Management and Introduce MixedPrecisionConfig (#328)
* refined to better support main params

* updated the use_ema and use_main_params flags to pass into parallelize, and get magifsdp policy kwargs

* fixed the is_first_fwd typo and copy_main_params record context

* inited magisdp_policy dataclass in magifsdp_config; passed magifsdp_policy to parallelize

* added specific properties for ckpt mgr to tell whether to load/save ema/main params

* minor fixed

* deprecated resume_config.load_ema_weights; added is_main_params_enabled and is_ema_params_enabled properties to MagiFSDPModule; removed use_main_params from base config

* added a [WIP] check for lint to early-stop ci if the PR is not really ready

* moved the check to dag_workflow and added [DO NOT MERGE]

* fixed ut for optim policy

* fixed the flags

* Refactor Main Param

* polish code and add ut

* add 2 functionality test

* Add new parallel plan impl

* Add mixed_precision_config

* fix ci

* fix ckpt save when use mamain param

* fix ci

* polish code

---------

+++ tests/test_functionality/test_fully_shard_comm.py
@@ -277,7 +277,7 @@ class TestFullyShardCollectiveOps(FSDPTestMultiThread):
             unsharded_grads,
             group,
             reduce_scatter_stream,
-            orig_dtype=orig_params[0].dtype,
+            cast_dtype=orig_params[0].dtype,
             reduce_dtype=reduce_scatter_dtype,
             device=self.device,
             reduce_scatter_reduce_op=None,
